backend/requirements.txt
yfinance==0.2.54
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.0
alpha_vantage==2.3.1 